v1.04対応

+++ install.sh
@@ -22,11 +22,14 @@ if [ -e ${User_servce} ]; then
     if [ "$enabled" = "enabled" ]; then
         sudo systemctl disable "$SERVICE_NAME"
     fi
+
+    if [ -e ${User_servce} ]; then
+        sudo rm -f /etc/systemd/system/dipper.service
+    fi
 fi
 
 # 以前のバージョンのアンインストール処理
 sudo rm -rf /usr/local/dipper/bin
-sudo rm -f /etc/systemd/system/dipper.service
 
 # v1.01以降のインストール用
 # スクリプトファイルダウンロード＆ファイル属性変更
